Flag explicitly certain arguments as of size_t type

+++ lib/Crypto/Util/strxor.py
@@ -32,7 +32,7 @@
 Fast XOR of byte strings.
 """
 
-from Crypto.Util._raw_api import (load_pycryptodome_raw_lib,
+from Crypto.Util._raw_api import (load_pycryptodome_raw_lib, c_size_t,
                                   create_string_buffer, get_raw_buffer)
 
 _raw_strxor = load_pycryptodome_raw_lib("Crypto.Util._strxor",
@@ -54,7 +54,7 @@ def strxor(term1, term2):
     if len(term1) != len(term2):
         raise ValueError("Only byte strings of equal length can be xored")
     result = create_string_buffer(len(term1))
-    _raw_strxor.strxor(term1, term2, result, len(term1))
+    _raw_strxor.strxor(term1, term2, result, c_size_t(len(term1)))
     return get_raw_buffer(result)
 
 
@@ -64,5 +64,5 @@ def strxor_c(term, c):
     if not 0 <= c < 256:
         raise ValueError("c must be in range(256)")
     result = create_string_buffer(len(term))
-    _raw_strxor.strxor_c(term, c, result, len(term))
+    _raw_strxor.strxor_c(term, c, result, c_size_t(len(term)))
     return get_raw_buffer(result)
